requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-unstructured
langchain-community
langchain-google-genai
langchain-huggingface
faiss-cpu
nltk
python-dotenv
streamlit
requests
sentence-transformers